+++ whyis/autonomic/global_change_service.py
@@ -5,7 +5,7 @@ import rdflib
 import setlr
 from datetime import datetime
 
-from autonomic.service import Service
+from .service import Service
 from nanopub import Nanopublication
 from datastore import create_id
 import flask
