+++ whyis/autonomic/ontology_importer.py
@@ -5,7 +5,7 @@ import rdflib
 import setlr
 from datetime import datetime
 
-from autonomic.update_change_service import UpdateChangeService
+from .update_change_service import UpdateChangeService
 from nanopub import Nanopublication
 from datastore import create_id
 import flask
